Refactor wrapper class placement and name
