+++ accomodation/src/main/java/com/codecool/accommodation/controller/SearchController.java
@@ -1,12 +1,8 @@
 package com.codecool.accommodation.controller;
 
-import com.codecool.accommodation.model.DTO.DTOWrapper;
-import com.codecool.accommodation.model.DTO.CoordinateDTO;
 import com.codecool.accommodation.service.SearchService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
